refactor: extract IDEFileManager from SikulixIDE (Phase 3, Step 1)
Move 18 file/session/image-parsing methods + 4 fields to IDEFileManager.java.
Delegation methods kept in SikulixIDE for external callers (ButtonCapture,
PatternWindow). Cleaned 6 unused imports from SikulixIDE.

SikulixIDE.java: 2665 -> 2344 lines (-321)
IDEFileManager.java: 411 lines (new)
